Bug-Fix: Close idle close_connections
Sessionmgr change to close all idle connections at
deletion time. This change also reduces the sessionmgr
ttl to 1 hour from the previously set 10hrs.
This is given that we have reduced the TTL on guitar
significantly also (1hr).

Verified this change with setting TTL to 5 min manually
https://screenshot.googleplex.com/74BrvTWcSoP8ypE
and ensuring connections are reclaimed.